requirements.txt
asgiref==3.3.1
astroid==2.5.1
autopep8==1.5.6
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.4
django-admin-bootstrapped==2.5.7
django-allauth==0.44.0
django-bootstrap-form==3.4
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-etc==1.3.0
django-filter==2.4.0
django-formtools==2.2
django-hitcount==1.3.3
django-jet==1.0.8
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
drf-registration==0.1.0
Faker==5.0.2
google-api-python-client==1.4.1
gunicorn==20.0.4
httplib2==0.9.2
idna==2.10
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
oauth2client==4.1.3
oauthlib==3.1.0
Pillow==8.1.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.0.1
pylint==2.7.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-http-client==3.3.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
sendgrid==6.5.0
six==1.15.0
source==1.2.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
stripe==2.56.0
text-unidecode==1.3
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
whitenoise==5.2.0
wrapt==1.12.1
